+++ notification-hubs-sample-app-kotlin/src/androidTest/java/com/example/notification_hubs_sample_app_kotlin/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.notification_hubs_test_app_kotlin
+package com.example.notification_hubs_sample_app_kotlin
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.notification_hubs_test_app_kotlin", appContext.packageName)
+        assertEquals("com.example.notification_hubs_sample_app_kotlin", appContext.packageName)
     }
 }
