+++ notification-hubs-sample-app-kotlin/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.notification_hubs_test_app_kotlin">
+    package="com.example.notification_hubs_sample_app_kotlin">
 
     <application
         android:allowBackup="true"
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="com.example.notification_hubs_test_app_kotlin.NotificationDetailActivity"></activity>
+        <activity android:name="com.example.notification_hubs_sample_app_kotlin.NotificationDetailActivity"></activity>
         <activity
-            android:name="com.example.notification_hubs_test_app_kotlin.MainActivity"
+            android:name="com.example.notification_hubs_sample_app_kotlin.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
