+++ notification-hubs-sample-app-kotlin/src/main/java/com/example/notification_hubs_sample_app_kotlin/MainActivity.kt
@@ -1,11 +1,11 @@
-package com.example.notification_hubs_test_app_kotlin
+package com.example.notification_hubs_sample_app_kotlin
 
 import android.os.Bundle
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.viewpager.widget.ViewPager
 
-import com.example.notification_hubs_test_app_kotlin.ui.main.SectionsPagerAdapter
+import com.example.notification_hubs_sample_app_kotlin.ui.main.SectionsPagerAdapter
 import com.google.android.material.tabs.TabLayout
 import com.microsoft.windowsazure.messaging.notificationhubs.InstallationTemplate
 import com.microsoft.windowsazure.messaging.notificationhubs.NotificationHub
@@ -21,7 +21,7 @@ class MainActivity : AppCompatActivity() {
         tabs.setupWithViewPager(viewPager)
 
         NotificationHub.start(this.application, BuildConfig.hubName, BuildConfig.hubListenConnectionString)
-        NotificationHub.addTag("userAgent:com.example.notification_hubs_test_app_kotlin:0.1.0")
+        NotificationHub.addTag("userAgent:com.example.notification_hubs_sample_app_kotlin:0.1.0")
         val testTemplate = InstallationTemplate()
         testTemplate.body = "{\"data\":{\"message\":\"Notification Hub test notification: \$myTextProp\"}}"
         NotificationHub.setTemplate("testTemplate", testTemplate)
