+++ notification-hubs-sample-app-kotlin/src/main/java/com/example/notification_hubs_sample_app_kotlin/cookbook/DevicePropertyProvider.kt
@@ -1,4 +1,4 @@
-package com.example.notification_hubs_test_app_kotlin.cookbook
+package com.example.notification_hubs_sample_app_kotlin.cookbook
 
 /**
  * Examples that are ready for copy/paste usage of how to get particular data about the device this
