+++ notification-hubs-sample-app-kotlin/src/main/java/com/example/notification_hubs_sample_app_kotlin/ui/main/PageViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.notification_hubs_test_app_kotlin.ui.main
+package com.example.notification_hubs_sample_app_kotlin.ui.main
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
